requirements.txt
psycopg2==2.8.3
requests==2.22.0
numpy==1.16.4
python-dateutil==2.8.0
matplotlib==3.1.0
pandas==0.24.2
gensim==3.7.3
nltk==3.4.1
spacy==2.1.4
tqdm==4.32.1
scipy==1.3.0
